feat: 修复bug、支持Linux跨平台编译

+++ Lite_version/sources/patch_kernel_root/analyze/kallsyms_lookup_name_4_6_0.cpp
@@ -54,11 +54,11 @@ bool KallsymsLookupName_4_6_0::init() {
 		std::cout << std::hex << "kallsyms_num: 0x" << m_kallsyms_num << ", offset: 0x" << kallsyms_num_offset << std::endl;
 
 		// revise the offset list offset again
-		const int offset_list_var_len = sizeof(long);
+		const int offset_list_var_len = sizeof(uint32_t);
 		offset_list_start = offset_list_end - m_kallsyms_num * offset_list_var_len;
-		long test_first_offset_list_val;
+		uint32_t test_first_offset_list_val;
 		do {
-			test_first_offset_list_val = *(long*)&m_file_buf[offset_list_start];
+			test_first_offset_list_val = *(uint32_t*)&m_file_buf[offset_list_start];
 			if (test_first_offset_list_val) {
 				offset_list_start -= offset_list_var_len;
 				offset_list_end -= offset_list_var_len;
@@ -252,18 +252,18 @@ bool KallsymsLookupName_4_6_0::find_kallsyms_addresses_list(std::vector<std::pai
 }
 
 static bool __find_kallsyms_offsets_list(const std::vector<char>& file_buf, size_t max_cnt, size_t& start, size_t& end) {
-	const int var_len = sizeof(long);
+	const int var_len = sizeof(uint32_t);
 	for (auto x = 0; x + var_len < file_buf.size(); x += var_len) {
-		long val1 = *(long*)&file_buf[x];
-		long val2 = *(long*)&file_buf[x + var_len];
+		uint32_t val1 = *(uint32_t*)&file_buf[x];
+		uint32_t val2 = *(uint32_t*)&file_buf[x + var_len];
 		if (val1 != 0 || val1 >= val2) {
 			continue;
 		}
 		int cnt = 0;
 		auto j = x + var_len;
 		for (; j + var_len < file_buf.size(); j += var_len) {
-			val1 = *(long*)&file_buf[j];
-			val2 = *(long*)&file_buf[j + var_len];
+			val1 = *(uint32_t*)&file_buf[j];
+			val2 = *(uint32_t*)&file_buf[j + var_len];
 			if (val1 > val2 || val2 == 0 || (val2 - val1) > 0x1000000) {
 				j += var_len;
 				break;
@@ -360,10 +360,10 @@ bool KallsymsLookupName_4_6_0::find_kallsyms_names_list(int kallsyms_num, size_t
 
 bool KallsymsLookupName_4_6_0::find_kallsyms_markers_list(int kallsyms_num, size_t name_list_end_offset, size_t& markers_list_start, size_t& markers_list_end) {
 	size_t start = align_up<8>(name_list_end_offset);
-	const int var_len = sizeof(long);
+	const int var_len = sizeof(uint32_t);
 	for (auto x = start; x + var_len < m_file_buf.size(); x += var_len) {
-		long val1 = *(long*)&m_file_buf[x];
-		long val2 = *(long*)&m_file_buf[x + var_len];
+		uint32_t val1 = *(uint32_t*)&m_file_buf[x];
+		uint32_t val2 = *(uint32_t*)&m_file_buf[x + var_len];
 		if (val1 == 0 && val2 > 0) {
 			markers_list_start = x;
 			break;
@@ -377,10 +377,10 @@ bool KallsymsLookupName_4_6_0::find_kallsyms_markers_list(int kallsyms_num, size
 
 	bool is_align8 = false;
 	int cnt = 5;
-	long last_second_var_val = 0;
+	uint32_t last_second_var_val = 0;
 	for (auto y = markers_list_start + var_len; y + var_len < m_file_buf.size(); y += var_len * 2) {
-		long val1 = *(long*)&m_file_buf[y];
-		long val2 = *(long*)&m_file_buf[y + var_len];
+		uint32_t val1 = *(uint32_t*)&m_file_buf[y];
+		uint32_t val2 = *(uint32_t*)&m_file_buf[y + var_len];
 		if (val2 != last_second_var_val) {
 			break;
 		}
@@ -398,9 +398,9 @@ bool KallsymsLookupName_4_6_0::find_kallsyms_markers_list(int kallsyms_num, size
 		} else {
 			markers_list_start -= back_val; // 4
 		}
-		markers_list_end = markers_list_start + ((kallsyms_num + 255) >> 8) * sizeof(long) * 2;
+		markers_list_end = markers_list_start + ((kallsyms_num + 255) >> 8) * sizeof(uint32_t) * 2;
 	} else {
-		markers_list_end = markers_list_start + ((kallsyms_num + 255) >> 8) * sizeof(long);
+		markers_list_end = markers_list_start + ((kallsyms_num + 255) >> 8) * sizeof(uint32_t);
 	}
 	
 	return true;
@@ -408,9 +408,9 @@ bool KallsymsLookupName_4_6_0::find_kallsyms_markers_list(int kallsyms_num, size
 
 bool KallsymsLookupName_4_6_0::find_kallsyms_token_table(size_t markers_list_end_offset, size_t& kallsyms_token_table_start, size_t& kallsyms_token_table_end) {
 	size_t start = align_up<8>(markers_list_end_offset);
-	const int var_len = sizeof(long);
+	const int var_len = sizeof(uint32_t);
 	for (auto x = start; x + var_len < m_file_buf.size(); x += var_len) {
-		long val1 = *(long*)&m_file_buf[x];
+		uint32_t val1 = *(uint32_t*)&m_file_buf[x];
 		if (val1 == 0) {
 			continue;
 		}
